Type the DevCom user word
`HSD_DevComRequest`'s last argument is declared `void* args`, but devcom hands
it to the completion callback as its second *integer* parameter
(`dc->callback(dc->dcReq, (int) dc->args, ...)`), and callers use it as a word
either way: `lbdvd` passes `(void*) index`, synth dressed a file offset up as
`(struct HSD_SynthStreamHeader*) (src + 0x20)`, `lbmemory` passes a real
pointer. Spell it `uintptr_t` — in the request, in the queue entry, in
`lbFile_800164A4`/`lbFile_80016580`, and as the callback typedef's second
parameter — so the seven casts inside devcom and the bogus struct cast in
synth are gone and the one genuine pointer keeps an explicit conversion.

Seven of the synth completion callbacks were declared with a different
parameter list and cast at the call site; they now have the signature the
typedef declares, and the `(HSD_DevComCallback) (Event)` double casts go.
`HSD_SynthPStreamFirstHakoDataCallback` keeps `(void)` and its cast: giving it
the four parameters costs four parameter homes and moves its frame.

`MCCPacket` moves to `hsd_3933.h` so the command table can be typed
`void (*)(MCCPacket*, MCCPacket*)`, which drops six casts on the packet
fields and two on the table's entries.

+++ src/sysdolphin/baselib/forward.h
@@ -110,7 +110,7 @@ typedef uintptr_t HSD_IDKey;
 
 typedef void (*GObj_RenderFunc)(HSD_GObj* gobj, int code);
 typedef void (*HSD_ObjUpdateFunc)(void* obj, enum_t type, HSD_ObjData* fval);
-typedef void (*HSD_DevComCallback)(int, int, void*, bool cancelflag);
+typedef void (*HSD_DevComCallback)(int, uintptr_t, void*, bool cancelflag);
 typedef void (*HSD_GObjEvent)(HSD_GObj* gobj);
 typedef void (*HSD_UserDataEvent)(void* user_data);
 typedef bool (*HSD_GObjPredicate)(HSD_GObj* gobj);
